+++ fields/TextHide/views/Cell.js
@@ -5,5 +5,5 @@ import {
 } from 'draftjs-conductor'
 
 export default function HtmlCell({ data }) {
-    return JSON.stringify(serialiseEditorStateToRaw(data))
+    return JSON.stringify(data)
 }